\name{.ellipsoid_data}
\alias{.ellipsoid_data}
\title{Calculates ellipsoid data for use in plot_cluster.}
\usage{.ellipsoid_data(data, x="x", y="y", cluster="cluster",
    ellipse_quantile=0.67)}
\description{Calculates ellipsoid data for use in plot_cluster.}
\keyword{internal}
\arguments{\item{data}{}
\item{x}{}
\item{y}{}
\item{cluster}{}
\item{ellipse_quantile}{}}
