\name{cluster_data.factanal}
\alias{cluster_data.factanal}
\title{Extracts primary principal components from object of type `factanal`.}
\usage{cluster_data.factanal(model, ...)}
\description{Extracts primary principal components from object of type `factanal`.}
\details{Generic function to extract cluster labels from a model into a data frame}
\seealso{\code{\link{cluster_data}}, \code{\link{plot_cluster}}}
\arguments{\item{model}{object of type factanal}
\item{...}{ignored}}
