\name{cluster_labels.tree}
\alias{cluster_labels.tree}
\title{Extract labels data frame from tree object for plotting using ggplot...}
\usage{cluster_labels.tree(model, ...)}
\description{Extract labels data frame from tree object for plotting using ggplot}
\details{Extract labels data frame from tree object for plotting using ggplot}
\value{a list with two elements: $labels and $leaf_labels}
\keyword{internal}
\seealso{\code{\link{cluster_data}}}
\arguments{\item{model}{object of class "tree", e.g. the output of tree()}
\item{...}{ignored}}
