\name{dendrogram_data}
\alias{dendrogram_data}
\title{Extract data frame from dendrogram object for plotting using ggplot...}
\usage{dendrogram_data(x, type=c("rectangle", "triangle"), ...)}
\description{Extract data frame from dendrogram object for plotting using ggplot}
\details{Extract data frame from dendrogram object for plotting using ggplot}
\seealso{\code{\link{cluster_data}}}
\keyword{internal}
\arguments{\item{x}{object of class "dendrogram", e.g. the output of as.dendrogram()}
\item{type}{The type of plot, indicating the shape of the dendrogram.  "Rectangle" will draw
rectangular lines, while "triangle" will draw triangular lines.}
\item{...}{ignored}}
