\name{get_cluster_data.kmeans}
\alias{get_cluster_data.kmeans}
\title{Extracts cluster allocation from object of type `kmeans`...}
\usage{get_cluster_data.kmeans(model, ...)}
\description{Extracts cluster allocation from object of type `kmeans`}
\details{Generic function to extract cluster labels from a model into a data frame}
\arguments{\item{model}{object of type kmeans}
\item{...}{ignored}}
