\name{get_cluster_data.Mclust}
\alias{get_cluster_data.Mclust}
\title{Extracts cluster allocation from object of type `Mclust`...}
\usage{get_cluster_data.Mclust(model, ...)}
\description{Extracts cluster allocation from object of type `Mclust`}
\details{Generic function to extract cluster labels from a model into a data frame}
\arguments{\item{model}{object of type kmeans}
\item{...}{ignored}}
