\name{get_cluster_data.princomp}
\alias{get_cluster_data.princomp}
\title{Extracts primary principal components from object of type `princomp`...}
\usage{get_cluster_data.princomp(model, ...)}
\description{Extracts primary principal components from object of type `princomp`}
\details{Generic function to extract cluster labels from a model into a data frame}
\arguments{\item{model}{object of type princomp}
\item{...}{ignored}}
