\name{get_cluster_labels.dendrogram}
\alias{get_cluster_labels.dendrogram}
\title{Extract data frame from dendrogram object for plotting using ggplot...}
\usage{get_cluster_labels.dendrogram(x, type=c("rectangle", "triangle"), ...)}
\description{Extract data frame from dendrogram object for plotting using ggplot}
\details{Extract data frame from dendrogram object for plotting using ggplot}
\seealso{\code{\link{get_cluster_data.dendrogram}}, 
\code{\link{get_cluster_labels.dendrogram}}}
\arguments{\item{x}{object of class "dendrogram", e.g. the output of as.dendrogram()}
\item{type}{The type of plot, indicating the shape of the dendrogram.  "Rectangle" will draw
rectangular lines, while "triangle" will draw triangular lines.}
\item{...}{ignored}}
