\name{get_cluster_labels.tree}
\alias{get_cluster_labels.tree}
\title{Extract labels data frame from tree object for plotting using ggplot...}
\usage{get_cluster_labels.tree(model, ...)}
\description{Extract labels data frame from tree object for plotting using ggplot}
\details{Extract labels data frame from tree object for plotting using ggplot}
\seealso{\code{\link{get_cluster_data}}}
\arguments{\item{model}{object of class "tree", e.g. the output of tree()}
\item{...}{ignored}}
