\name{pca_fit_loadings}
\alias{pca_fit_loadings}
\title{Calculate pca fit loadings...}
\usage{pca_fit_loadings(x, labels_at_edge, explode)}
\description{Calculate pca fit loadings}
\keyword{Internal}
\arguments{\item{x}{Cluster model, e.g. the results of kmeans}
\item{labels_at_edge}{Indicates whether labels are at edge of graph or next to poit}
\item{explode}{Explode multiplication factor}}
