\name{cluster_data.kmeans}
\alias{cluster_data.kmeans}
\title{Extracts cluster allocation from object of type `kmeans`...}
\usage{cluster_data.kmeans(model, ...)}
\description{Extracts cluster allocation from object of type `kmeans`}
\details{Generic function to extract cluster labels from a model into a data frame}
\seealso{\code{\link{plot_cluster}}}
\arguments{\item{model}{object of type kmeans}
\item{...}{ignored}}
