\name{cluster_data.Mclust}
\alias{cluster_data.Mclust}
\title{Extracts cluster allocation from object of type `Mclust`...}
\usage{cluster_data.Mclust(model, ...)}
\description{Extracts cluster allocation from object of type `Mclust`}
\details{Generic function to extract cluster labels from a model into a data frame}
\seealso{\code{\link{plot_cluster}}}
\arguments{\item{model}{object of type Mclust}
\item{...}{ignored}}
