\name{cluster_data.prcomp}
\alias{cluster_data.prcomp}
\title{Extracts primary principal components from object of type `prcomp`.}
\usage{cluster_data.prcomp(model, ...)}
\description{Extracts primary principal components from object of type `prcomp`.}
\details{Generic function to extract cluster labels from a model into a data frame}
\seealso{\code{\link{cluster_data}}, \code{\link{plot_cluster}}}
\arguments{\item{model}{object of type prcomp}
\item{...}{ignored}}
