\name{cluster_data.princomp}
\alias{cluster_data.princomp}
\title{Extracts primary principal components from object of type `princomp`.}
\usage{cluster_data.princomp(model, ...)}
\description{Extracts primary principal components from object of type `princomp`.}
\details{Generic function to extract cluster labels from a model into a data frame}
\seealso{\code{\link{cluster_data}}, \code{\link{plot_cluster}}}
\arguments{\item{model}{object of type princomp}
\item{...}{ignored}}
