\name{get_cluster_data.kmeans}
\alias{get_cluster_data.kmeans}
\title{Generic function to extract cluster labels from a model into a data frame...}
\usage{get_cluster_data.kmeans(model, ...)}
\description{Generic function to extract cluster labels from a model into a data frame}
\details{Generic function to extract cluster labels from a model into a data frame}
\arguments{\item{model}{object of type hclust, dendrogram, tree or kmeans}
\item{...}{ignored}}
